+++ includes/admin/class-cocart-admin-assets.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * CoCart - Admin Assets.
+ * Manages CoCart dashboard assets.
  *
  * @since    1.2.0
  * @version  2.0.3