+++ includes/admin/class-cocart-admin.php
@@ -35,13 +35,16 @@ if ( ! class_exists( 'CoCart_Admin' ) ) {
 		/**
 		 * Include any classes we need within admin.
 		 *
-		 * @access public
+		 * @access  public
+		 * @since   1.2.0
+		 * @version 2.1.0
 		 */
 		public function includes() {
-			include( dirname( __FILE__ ) . '/class-cocart-admin-action-links.php' );     // Action Links
-			include( dirname( __FILE__ ) . '/class-cocart-admin-assets.php' );           // Admin Assets
-			include( dirname( __FILE__ ) . '/class-cocart-admin-notices.php' );          // Plugin Notices
-			include( dirname( __FILE__ ) . '/class-cocart-wc-admin-system-status.php' ); // WooCommerce System Status
+			include( dirname( __FILE__ ) . '/class-cocart-admin-action-links.php' );         // Action Links
+			include( dirname( __FILE__ ) . '/class-cocart-admin-assets.php' );               // Admin Assets
+			include( dirname( __FILE__ ) . '/class-cocart-admin-plugin-screen-update.php' ); // Plugin Screen Update
+			include( dirname( __FILE__ ) . '/class-cocart-admin-notices.php' );              // Plugin Notices
+			include( dirname( __FILE__ ) . '/class-cocart-wc-admin-system-status.php' );     // WooCommerce System Status
 		} // END includes()
 
 		/**