+++ includes/admin/views/html-notice-requirement-wp.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Admin View: WordPress Requirment Notice.
+ * Admin View: WordPress Requirement Notice.
  *
  * @since    1.2.0
  * @version  2.0.0