+++ includes/admin/views/html-notice-upgrade-warning.php
@@ -3,7 +3,7 @@
  * Admin View: Upgrade Warning Notice.
  *
  * @since    1.2.3
- * @version  2.0.9
+ * @version  2.0.11
  * @author   Sébastien Dumont
  * @category Admin
  * @package  CoCart
@@ -18,17 +18,17 @@ if ( ! defined( 'ABSPATH' ) ) {
 <div class="notice notice-info cocart-notice">
 	<div class="cocart-notice-inner">
 		<div class="cocart-notice-icon">
-			<img src="<?php echo COCART_URL_PATH . '/assets/images/logo.jpg'; ?>" alt="<?php echo esc_attr__( 'CoCart, a WooCommerce REST-API extension', 'cart-rest-api-for-woocommerce' ); ?>" />
+			<img src="<?php echo COCART_URL_PATH . '/assets/images/logo.jpg'; ?>" alt="CoCart Logo" />
 		</div>
 
 		<div class="cocart-notice-content">
 			<h3><?php esc_html_e( 'Upgrade Warning!', 'cart-rest-api-for-woocommerce' ); ?></h3>
-			<p><?php echo sprintf( __( 'Version %1$s%4$s%2$s of %3$s is coming soon and will provide support for guest customers among new filters for developers. Before it is released, I require your help to test it out and provide your feedback! Documentation is also available for these features.', 'cart-rest-api-for-woocommerce' ), '<strong>', '</strong>', 'CoCart', COCART_NEXT_VERSION ); ?></p>
+			<p><?php echo sprintf( __( 'Version %1$s%4$s%2$s of %3$s is coming soon and will provide support for guest customers among new filters for developers. I am in need of testers and your feedback.', 'cart-rest-api-for-woocommerce' ), '<strong>', '</strong>', 'CoCart', COCART_NEXT_VERSION ); ?></p>
 		</div>
 
 		<div class="cocart-action">
-			<?php printf( '<a href="%1$s" class="button button-primary cocart-button" target="_blank">%2$s</a>', esc_url( COCART_NEXT_VERSION_DETAILS ), esc_html__( 'Learn More', 'cart-rest-api-for-woocommerce' ) ); ?>
-			<a href="<?php echo esc_url( add_query_arg( 'hide_cocart_upgrade_notice', 'true' ) ); ?>" class="no-thanks"><?php echo esc_html__( 'Don\'t ask me again for another 2 weeks', 'cart-rest-api-for-woocommerce' ); ?></a>
+			<?php printf( '<a href="%1$s" class="button button-primary cocart-button" target="_blank">%2$s</a>', esc_url( 'https://cocart.xyz/contact/' ), esc_html__( 'Sign up to Test', 'cart-rest-api-for-woocommerce' ) ); ?>
+			<a href="<?php echo esc_url( add_query_arg( 'hide_cocart_upgrade_notice', 'true' ) ); ?>" class="no-thanks"><?php echo esc_html__( 'Ask me again in 2 weeks', 'cart-rest-api-for-woocommerce' ); ?></a>
 		</div>
 	</div>
 </div>