+++ CHANGELOG.md
@@ -30,6 +30,21 @@
 > This release is a work in progress and requires developer feedback.
 > Developers please see documentation for more details on the new hooks added.
 
+## v2.0.12 - 27th March, 2020
+
+* Tested: Compatible with WordPress 5.4
+* Added: Upgrade notice message on plugins upgrade screen.
+
+## v2.0.11 - 25th March, 2020
+
+* Removed: `cocart_docs_url` filter for changing documentation link.
+* Tested: Compatible with WooCommerce v4.0.x
+* Updated: Getting Started page and removed `cocart_getting_started_doc_url` filter for the documentation button.
+* Updated: Plugin action link for upgrading to CoCart Pro.
+* Updated: Upgrade notices.
+
+> Please temporarily deactivate CoCart and CoCart Pro if you have it before updating WooCommerce to version 4.0+ as there is an activation order issue I am still working on fixing. Once you have upgraded WooCommerce simply reactivate CoCart.
+
 ## v2.0.10 - 22nd March, 2020
 
 * Tweaked: Refresh totals parameter is now set to `true` by default when item is updated.